+++ android/src/main/java/com/reactnativestripesdk/CardFieldViewManager.kt
@@ -5,10 +5,9 @@ import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.common.MapBuilder
 import com.facebook.react.uimanager.SimpleViewManager
 import com.facebook.react.uimanager.ThemedReactContext
-import com.facebook.react.uimanager.UIManagerModule
 import com.facebook.react.uimanager.annotations.ReactProp
 
-class StripeSdkCardViewManager : SimpleViewManager<StripeSdkCardView>() {
+class CardFieldViewManager : SimpleViewManager<CardFieldView>() {
   override fun getName() = "CardField"
 
   private var reactContextRef: ThemedReactContext? = null
@@ -19,7 +18,7 @@ class StripeSdkCardViewManager : SimpleViewManager<StripeSdkCardView>() {
       CardChangedEvent.EVENT_NAME, MapBuilder.of("registrationName", "onCardChange"))
   }
 
-  override fun receiveCommand(root: StripeSdkCardView, commandId: String?, args: ReadableArray?) {
+  override fun receiveCommand(root: CardFieldView, commandId: String?, args: ReadableArray?) {
     when (commandId) {
       "focus" -> root.requestFocusFromJS()
       "blur" -> root.requestBlurFromJS()
@@ -28,33 +27,38 @@ class StripeSdkCardViewManager : SimpleViewManager<StripeSdkCardView>() {
   }
 
   @ReactProp(name = "dangerouslyGetFullCardDetails")
-  fun setDangerouslyGetFullCardDetails(view: StripeSdkCardView, dangerouslyGetFullCardDetails: Boolean = false) {
-    view.setDangerouslyGetFullCardDetails(dangerouslyGetFullCardDetails);
+  @SuppressWarnings("unused")
+  fun setDangerouslyGetFullCardDetails(view: CardFieldView, dangerouslyGetFullCardDetails: Boolean = false) {
+    view.setDangerouslyGetFullCardDetails(dangerouslyGetFullCardDetails)
   }
 
   @ReactProp(name = "postalCodeEnabled")
-  fun setPostalCodeEnabled(view: StripeSdkCardView, postalCodeEnabled: Boolean = true) {
-    view.setPostalCodeEnabled(postalCodeEnabled);
+  @SuppressWarnings("unused")
+  fun setPostalCodeEnabled(view: CardFieldView, postalCodeEnabled: Boolean = true) {
+    view.setPostalCodeEnabled(postalCodeEnabled)
   }
 
   @ReactProp(name = "autofocus")
-  fun setAutofocus(view: StripeSdkCardView, autofocus: Boolean = false) {
-    view.setAutofocus(autofocus);
+  @SuppressWarnings("unused")
+  fun setAutofocus(view: CardFieldView, autofocus: Boolean = false) {
+    view.setAutofocus(autofocus)
   }
 
   @ReactProp(name = "cardStyle")
-  fun setCardStyle(view: StripeSdkCardView, cardStyle: ReadableMap) {
-    view.setCardStyle(cardStyle);
+  @SuppressWarnings("unused")
+  fun setCardStyle(view: CardFieldView, cardStyle: ReadableMap) {
+    view.setCardStyle(cardStyle)
   }
 
   @ReactProp(name = "placeholder")
-  fun setPlaceHolders(view: StripeSdkCardView, placeholder: ReadableMap) {
-    view.setPlaceHolders(placeholder);
+  @SuppressWarnings("unused")
+  fun setPlaceHolders(view: CardFieldView, placeholder: ReadableMap) {
+    view.setPlaceHolders(placeholder)
   }
 
-  override fun createViewInstance(reactContext: ThemedReactContext): StripeSdkCardView {
+  override fun createViewInstance(reactContext: ThemedReactContext): CardFieldView {
     val stripeSdkModule: StripeSdkModule? = reactContext.getNativeModule(StripeSdkModule::class.java)
-    val view = StripeSdkCardView(reactContext)
+    val view = CardFieldView(reactContext)
 
     reactContextRef = reactContext
 
@@ -62,7 +66,7 @@ class StripeSdkCardViewManager : SimpleViewManager<StripeSdkCardView>() {
     return view
   }
 
-  override fun onDropViewInstance(view: StripeSdkCardView) {
+  override fun onDropViewInstance(view: CardFieldView) {
     super.onDropViewInstance(view)
 
     val stripeSdkModule: StripeSdkModule? = reactContextRef?.getNativeModule(StripeSdkModule::class.java)