+++ android/src/main/java/com/reactnativestripesdk/FormCompleteEvent.kt
@@ -1,6 +1,5 @@
 package com.reactnativestripesdk
 import com.facebook.react.bridge.Arguments
-import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.bridge.WritableMap
 import com.facebook.react.uimanager.events.Event
 import com.facebook.react.uimanager.events.RCTEventEmitter