+++ ios/StripeSdk.m
@@ -70,7 +70,7 @@ RCT_EXTERN_METHOD(
                   )
 
 RCT_EXTERN_METHOD(
-                  handleCardAction:(NSString *)paymentIntentClientSecret
+                  handleNextAction:(NSString *)paymentIntentClientSecret
                   resolver: (RCTPromiseResolveBlock)resolve
                   rejecter: (RCTPromiseRejectBlock)reject
                   )